removed configureStore.prod & configureStore.dev files
