+++ app/store/configureStore.js
@@ -1,5 +1,23 @@
+import { createStore, applyMiddleware, compose } from "redux";
+import thunk from 'redux-thunk';
+import rootReducer from "./reducers";
+import logger from 'redux-logger';
+import DevTools from "../containers/DevTools";
+
+let enhancer = null;
+
 if (process.env.NODE_ENV === "production") {
-  module.exports = require("./configureStore.prod");
+    enhancer = compose(
+        applyMiddleware(thunk)
+    );
 } else {
-  module.exports = require("./configureStore.dev");
+    enhancer = compose(
+        applyMiddleware(thunk, logger),
+        DevTools.instrument()
+    );
+}
+
+export default function configureStore(initialState) {
+    const store = createStore(rootReducer, initialState, enhancer);
+    return store;
 }
